+++ lib/src/validators/async_validator.dart
@@ -4,9 +4,11 @@
 
 import 'package:reactive_forms/reactive_forms.dart';
 
+/// An abstract class extended by classes that perform asynchronous validation.
 abstract class AsyncValidator<T> {
   const AsyncValidator();
 
+  /// Validates the [control].
   Future<Map<String, dynamic>?> validate(AbstractControl<T> control);
 
   Future<Map<String, dynamic>?> call(AbstractControl<T> control) {