+++ lib/src/validators/must_match_validator.dart
@@ -13,7 +13,8 @@ class MustMatchValidator extends Validator<dynamic> {
 
   /// Constructs an instance of [MustMatchValidator]
   const MustMatchValidator(
-      this.controlName, this.matchingControlName, this.markAsDirty) : super();
+      this.controlName, this.matchingControlName, this.markAsDirty)
+      : super();
 
   @override
   Map<String, dynamic>? validate(AbstractControl<dynamic> control) {