+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 ## Breaking Changes
 
-- All validators have been changed to classes.
-- A new validator `DelegateValidator` has been introduced to be used with a custom validation function.
+- All validators have been changed to classes with `const` constructors.
+- The Asynchronous Validator is now a class from where any custom async validator can inherit.
+
+## Features
+
+- A new validator `Validators.delegate(...)` has been introduced to be used with a custom
+  validation function.
+- A new validator `Validators.delegateAsync(...)` has been introduced to be used with a custom
+  async validation functions.
 
 # 14.3.0